% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beta_all}
\alias{beta_all}
\title{Mean nearest diversity within a distance matrix}
\usage{
beta_all(comm, traits, coph)
}
\arguments{
\item{comm}{a community matrix with sites as rows and species as columns}

\item{coph}{a dataframe of trait values with species rownames}

\item{coph}{a species by species cophenetic matrix}
}
\description{
Computes mean nearest neighbor distance for a pair of cells
}
\seealso{
picante package for mulitple methods (phylosor etc.) picante::cophenetic
}

