% Generated by roxygen2 (4.0.2): do not edit by hand
\name{betaPar.scatter}
\alias{betaPar.scatter}
\title{Betadiversity metrics for list of indices}
\usage{
betaPar.scatter(toScatterMatrix, toScatterIndex, coph, traits)
}
\arguments{
\item{toScatterMatrix}{a community matrix with sites as rows and species as columns}

\item{toScatterIndex}{an 2 row by n columns index of rows in comm to subset}

\item{coph}{a species by species cophenetic matrix}

\item{traits}{a species by species cophenetic matrix}
}
\description{
Computes taxonomic, phylogenetic and trait metrics for each pair of cells
}

