% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MNND}
\alias{MNND}
\title{Mean nearest diversity within a distance matrix}
\usage{
MNND(A, B, sp.list, dists)
}
\arguments{
\item{A}{Assemblage 1, matching name in assemblage list}

\item{B}{Assemblage 2, matching name in assemblage list}

\item{dists}{a species by species matrix of distances}

\item{sp.lists}{a species list at each cell, named for each assemblage}
}
\description{
Computes mean nearest neighbor distance for a pair of cells
}
\seealso{
picante package for mulitple methods (phylosor etc.) picante::cophenetic
}

