% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MNTDt}
\alias{MNTDt}
\title{Phylogenetic diversity between cells}
\usage{
MNTDt(comm, dists, sp.lists, nam = "value")
}
\arguments{
\item{comm}{a community matrix with sites as rows and species as columns}

\item{sp.lists}{a species list at each cell}

\item{nam}{a column name for the distance matrix, defaults to value}

\item{coph}{a species by species cophenetic matrix}
}
\description{
Computes mean nearest neighbor distance for a pair of cells
}
\seealso{
picante package for mulitple methods (phylosor etc.) picante::cophenetic
}

