% Generated by roxygen2 (4.0.2): do not edit by hand
\name{phyloMNTDt}
\alias{phyloMNTDt}
\title{Phylogenetic diversity between cells}
\usage{
phyloMNTDt(comm, coph, sp.lists)
}
\arguments{
\item{comm}{a community matrix with sites as rows and species as columns}

\item{coph}{a species by species cophenetic matrix}

\item{sp.lists}{a species list at each cell}
}
\description{
Computes mean nearest neighbor distance for a pair of cells
}
\seealso{
picante package for mulitple methods (phylosor etc.) picante::cophenetic
}

