% Generated by roxygen2 (4.0.2): do not edit by hand
\name{taxF}
\alias{taxF}
\title{Taxonomic diversity between cells}
\usage{
taxF(comm)
}
\arguments{
\item{comm}{a community matrix with sites as rows and species as columns}
}
\description{
Computes betasim for a pair of cells
}
\seealso{
beta.pairs in the betapart packages
}

