% Generated by roxygen2 (4.0.2): do not edit by hand
\name{traitPCA}
\alias{traitPCA}
\title{Calculate trait space}
\usage{
traitPCA(comm, traits)
}
\arguments{
\item{comm}{a community matrix with sites as rows and species as columns}

\item{traits}{a species by trait matrix of trait values}
}
\description{
Computes trait loadings for species in each assemblage
}
\seealso{
MNTDt
}

