% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matpsim}
\alias{matpsim}
\title{Cell by branch matrix calculation}
\usage{
matpsim(com, tcellbr)
}
\arguments{
\item{com}{a community matrix with sites as rows and species as columns}

\item{phyl}{a phylo object: phylogenetic tree holding species in the community}
}
\description{
This function computes the presence of species in each cell by evaluating which nodes are present. Often used for phylogenetic community structure analysis
}
\seealso{
picante::mpd
}

